load table double to decimal

+++ ElsterA1140Reader/LoadTable.cs
@@ -10,9 +10,9 @@ namespace ElsterA1140Reader
     {
         public DateTime? From { get; set; }
         public DateTime? To { get; set; }
-        public double Cn1 { get; set; }
-        public double Cn2 { get; set; }
-        public double Cn3 { get; set; }
-        public double Cn4 { get; set; }
+        public decimal Cn1 { get; set; }
+        public decimal Cn2 { get; set; }
+        public decimal Cn3 { get; set; }
+        public decimal Cn4 { get; set; }
     }
 }
